+++ js/ai/src/index.ts
@@ -61,6 +61,7 @@ export {
 } from './generate.js';
 export { Message } from './message.js';
 export {
+  GenerateActionOptionsSchema,
   GenerateResponseChunkSchema,
   GenerationCommonConfigSchema,
   MessageSchema,
